requirements.txt
flask
pandas
folium
paho-mqtt
